fix(build): complete plugin scaffold metadata
