+++ mill-plugins/morphir/core/src/org/finos/morphir/mill/PluginPackage.scala
@@ -1,0 +1,3 @@
+package org.finos.morphir.mill
+
+private[mill] object PluginPackage
